Getting back to working

+++ src/map.hpp
@@ -62,8 +62,9 @@ struct Room {
     u8 side_id;
     ScrollType scroll;
     // X, Y position of the top left corner of the whole room.
-    u8 x;
-    u8 y;
+    // Stored as a 16 bit number just to make comparisons between player position faster
+    s16 x;
+    s16 y;
 
     // The object type that drops when this room is cleared
     // Update this if the prize has been grabbed
